Completes Tier 1 of the fable2 review (points 1, 3, 4, 5; point 2 already
merged). Seven commits:
  - T1.3 extract PickTarget + ShouldHandoffFizzledPull pure kernels + golden tests
  - T1.5 DcPullDecisionCode enum + camp-ownership helpers + DcRel relevance ladder
  - T1.1 event registry cross-reference + persistence lint gtests
  - T1.4 drop dead Config.h includes (partial IWYU)
  - 3 compile fixups (campUnset capture, apply() enum call sites)

Build-verified by the user.

Session: S869

+++ src/Ai/Dungeon/DungeonClear/Util/DcStatusPublisher.cpp
@@ -26,7 +26,6 @@
 #include <vector>
 #include "AttackersValue.h"
 #include "CellImpl.h"
-#include "Config.h"
 #include "Creature.h"
 #include "CreatureGroups.h"
 #include "GameObject.h"
@@ -218,7 +217,7 @@ std::string DcStatusPublisher::BuildStatusPayload(PlayerbotAI* botAI)
     bool const pullMode = AI_VALUE(bool, "dungeon clear pull mode");
     uint32 const pullSetting = AI_VALUE(uint32, "dungeon clear pull setting");
     DcPullContext const& pull = AI_VALUE(DcPullContext&, "dungeon clear pull context");
-    uint32 const pullDecision = pull.decision;
+    uint32 const pullDecision = static_cast<uint32>(pull.decision);
     uint32 const pullPhase = static_cast<uint32>(pull.phase);
     std::string const bossName = next.has_value() ? next->name : "the boss";
